refactor: 요구사항 충족을 위한 디렉터리명 수정
도메인의 의미를 명시하기 위해 폴더명을 domain으로 수정
