+++ src/domain/Winner.js
